requirements.txt
flask
biopython
pandas
requests
matplotlib
numpy
gunicorn